Update README.md
Have a socket with a server and client side. Randomly arrange a 52 card deck and send it over to the player side giving them a choice to hit/stay and also have a card counting recommendation that will count the face cards and either move up the count or down depending on the cards at hand.